GumikView-on lehet kosárba rakni a gumikból maximum rendelhető a raktárban lévő darabszám

+++ src/main/java/panisz/norbert/simongumis/entities/RendelesiEgysegEntity.java
@@ -13,5 +13,5 @@ public class RendelesiEgysegEntity extends BaseEntity {
     @OneToOne
     private GumikEntity gumi;
     private Integer mennyiseg;
-    private Float reszosszeg;
+    private Integer reszosszeg;
 }
